requirements.txt
asgiref==3.5.0
Django==4.0.4
django-colorfield==0.6.3
django-debug-toolbar==3.2.4
django-environ==0.8.1
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
